+++ src/Trading.Bot/Configuration/TradeSettings.cs
@@ -7,9 +7,9 @@ public class TradeSettings
     public string[] OtherGranularities { get; set; }
     public TimeSpan CandleSpan { get; set; }
     public int[] Integers { get; set; }
-    public double[] Doubles { get; set; }
-    public double MaxSpread { get; set; }
-    public double MinGain { get; set; }
-    public double RiskReward { get; set; }
+    public decimal[] Doubles { get; set; }
+    public decimal MaxSpread { get; set; }
+    public decimal MinGain { get; set; }
+    public decimal RiskReward { get; set; }
     public bool TrailingStop { get; set; }
 }
