+++ src/Trading.Bot/Models/ApiResponses/OrderFilledResponse.cs
@@ -2,28 +2,28 @@
 
 public class OrderFilledResponse
 {
-    public double AccountBalance { get; set; }
+    public decimal AccountBalance { get; set; }
     public string AccountID { get; set; }
     public string BatchID { get; set; }
     public string RequestID { get; set; }
-    public double Commission { get; set; }
-    public double Financing { get; set; }
-    public double FullVWAP { get; set; }
+    public decimal Commission { get; set; }
+    public decimal Financing { get; set; }
+    public decimal FullVWAP { get; set; }
     public FullPrice FullPrice { get; set; }
-    public double GuaranteedExecutionFee { get; set; }
-    public double QuoteGuaranteedExecutionFee { get; set; }
-    public double HalfSpreadCost { get; set; }
+    public decimal GuaranteedExecutionFee { get; set; }
+    public decimal QuoteGuaranteedExecutionFee { get; set; }
+    public decimal HalfSpreadCost { get; set; }
     public string Id { get; set; }
     public string Instrument { get; set; }
     public string OrderID { get; set; }
-    public double Pl { get; set; }
-    public double QuotePl { get; set; }
+    public decimal Pl { get; set; }
+    public decimal QuotePl { get; set; }
     public string Reason { get; set; }
     public DateTime Time { get; set; }
     public TradeOpened TradeOpened { get; set; }
     public TradeClosed[] TradesClosed { get; set; }
     public string Type { get; set; }
-    public double Units { get; set; }
+    public decimal Units { get; set; }
     public int UserID { get; set; }
 }
 
@@ -31,21 +31,21 @@ public class FullPrice
 {
     public PriceResponse.Ask[] Asks { get; set; }
     public PriceResponse.Bid[] Bids { get; set; }
-    public double CloseoutAsk { get; set; }
-    public double CloseoutBid { get; set; }
+    public decimal CloseoutAsk { get; set; }
+    public decimal CloseoutBid { get; set; }
     public DateTime Timestamp { get; set; }
 }
 
 public class TradeOpened
 {
     public string TradeID { get; set; }
-    public double Units { get; set; }
-    public double Price { get; set; }
+    public decimal Units { get; set; }
+    public decimal Price { get; set; }
 }
 
 public class TradeClosed
 {
     public string TradeID { get; set; }
-    public double Units { get; set; }
-    public double Price { get; set; }
+    public decimal Units { get; set; }
+    public decimal Price { get; set; }
 }
