+++ src/Trading.Bot/Models/ApiResponses/TradeResponse.cs
@@ -2,17 +2,17 @@
 
 public class TradeResponse
 {
-    public double CurrentUnits { get; set; }
-    public double Financing { get; set; }
+    public decimal CurrentUnits { get; set; }
+    public decimal Financing { get; set; }
     public string Id { get; set; }
-    public double InitialUnits { get; set; }
+    public decimal InitialUnits { get; set; }
     public string Instrument { get; set; }
     public DateTime OpenTime { get; set; }
-    public double Price { get; set; }
-    public double RealizedPL { get; set; }
+    public decimal Price { get; set; }
+    public decimal RealizedPL { get; set; }
     public string State { get; set; }
-    public double UnrealizedPL { get; set; }
-    public double MarginUsed { get; set; }
+    public decimal UnrealizedPL { get; set; }
+    public decimal MarginUsed { get; set; }
     public ClientExtensions ClientExtensions { get; set; }
     public TakeProfit TakeProfitOrder { get; set; }
     public StopLoss StopLossOrder { get; set; }
